Updated the code so that all standard code checks that should be applied when releasing a module to CPAN run fine.
Most of the changes are formatting changes.
I added documentation stubs for all member functions of the module. The documentation should still be reviewed and completed, preferrably by the original author :-)
One functional change: passing the IP address of the bridge is enough, the 'http://' prefix is prepended if it is not present.